@@ -143,7 +143,7 @@
     "react-native": "local-cli/wrong-react-native.js"
   },
   "peerDependencies": {
-    "react": "16.2.0"
+    "react": "^16.3.0-alpha.0"
   },
   "dependencies": {
     "absolute-path": "^0.0.0",
@@ -219,8 +219,8 @@
     "jest": "22.2.1",
     "jest-junit": "3.5.0",
     "prettier": "1.9.1",
-    "react": "16.2.0",
-    "react-test-renderer": "16.2.0",
+    "react": "^16.3.0-alpha.0",
+    "react-test-renderer": "^16.3.0-alpha.0",
     "shelljs": "^0.7.8",
     "sinon": "^2.2.0"
   }
