@@ -41,7 +41,7 @@ var ListViewSimpleExample = createReactClass({
 
   _pressData: ({}: {[key: number]: boolean}),
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._pressData = {};
   },
 
