@@ -42,7 +42,7 @@ var SwipeableListViewSimpleExample = createReactClass({
 
   _pressData: ({}: {[key: number]: boolean}),
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._pressData = {};
   },
 
