@@ -334,7 +334,7 @@ class AutogrowingTextInputExample extends React.Component<{}> {
     };
   }
 
-  componentWillReceiveProps(props) {
+  UNSAFE_componentWillReceiveProps(props) {
     this.setState({
       multiline: props.multiline,
     });
