@@ -1,3 +1,12 @@
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
 'use strict';
 
 /* @flow */
@@ -58,7 +67,7 @@ class KeyboardSpacerIOS extends Component<Props, State> {
     keyboardHeight: 0,
   };
 
-  componentWillMount() {
+  UNSAFE_componentWillMount() {
     this._registerEvents();
   }
 
